Add config for test db, reorganize tests by module
